Добавление проверки на количество букетов на главной странице

+++ src/main/java/com/example/OliviaFlowers/controllers/HomePageController.java
@@ -10,6 +10,7 @@ import org.springframework.web.bind.annotation.GetMapping;
 import com.example.OliviaFlowers.secvices.BouquetService;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 import java.util.List;
 
@@ -38,17 +39,18 @@ public class HomePageController {
             // Поместить полученные букеты в модель
             model.addAttribute("homePageBouquets", homePageBouquets);
 
+            if (homePageBouquets.size() == 3) {
+                Bouquet bouquet1 = homePageBouquets.get(0).getBouquet();
+                Bouquet bouquet2 = homePageBouquets.get(1).getBouquet();
+                Bouquet bouquet3 = homePageBouquets.get(2).getBouquet();
 
-            // Получить информацию о каждом букете и добавить её в модель
-            Bouquet bouquet1 = homePageBouquets.get(0).getBouquet();
-            Bouquet bouquet2 = homePageBouquets.get(1).getBouquet();
-            Bouquet bouquet3 = homePageBouquets.get(2).getBouquet();
+                model.addAttribute("bouquet1", bouquet1);
+                model.addAttribute("bouquet2", bouquet2);
+                model.addAttribute("bouquet3", bouquet3);
+            } else {
+                // Логика для обработки случая, когда список пустой
+            }
 
-            model.addAttribute("bouquet1", bouquet1);
-            model.addAttribute("bouquet2", bouquet2);
-            model.addAttribute("bouquet3", bouquet3);
-
-            model.addAttribute("homePageBouquets", homePageBouquets);
         } catch (Exception e) {
             e.printStackTrace();
         }
@@ -58,36 +60,42 @@ public class HomePageController {
 
 
     @PostMapping("/add_bouquets_to_homepage")
-    public String addBouquetsToHomepage(@RequestParam(name = "bouquet1") Long bouquetId1, @RequestParam(name = "description1") String description1,
-                                        @RequestParam(name = "bouquet2") Long bouquetId2, @RequestParam(name = "description2") String description2,
-                                        @RequestParam(name = "bouquet3") Long bouquetId3, @RequestParam(name = "description3") String description3,
-                                        Model model) {
+    public String addBouquetsToHomepage(@RequestParam(name = "bouquet1", required = false) Long bouquetId1, @RequestParam(name = "description1", required = false) String description1,
+                                        @RequestParam(name = "bouquet2", required = false) Long bouquetId2, @RequestParam(name = "description2", required = false) String description2,
+                                        @RequestParam(name = "bouquet3", required = false) Long bouquetId3, @RequestParam(name = "description3", required = false) String description3,
+                                        Model model, RedirectAttributes redirectAttributes) {
+        if (bouquetId1 != null && bouquetId2 != null && bouquetId3 != null
+            && description1 != null && description2 != null && description3 != null) {
+            homePageRepository.deleteAll();
 
-        homePageRepository.deleteAll();
-
-        // Получить информацию о букетах с переданными ID и добавить её в модель
-        Bouquet bouquet1 = bouquetService.getBouquetByID(bouquetId1);
-        Bouquet bouquet2 = bouquetService.getBouquetByID(bouquetId2);
-        Bouquet bouquet3 = bouquetService.getBouquetByID(bouquetId3);
+            // Получить информацию о букетах с переданными ID и добавить её в модель
+            Bouquet bouquet1 = bouquetService.getBouquetByID(bouquetId1);
+            Bouquet bouquet2 = bouquetService.getBouquetByID(bouquetId2);
+            Bouquet bouquet3 = bouquetService.getBouquetByID(bouquetId3);
 
 
-        // Создать новую запись для главной страницы и сохранить её в базе данных
-        HomePage homePage1 = new HomePage();
-        homePage1.setBouquet(bouquet1);
-        homePage1.setDescription(description1);
-        homePageRepository.save(homePage1);
+            // Создать новую запись для главной страницы и сохранить её в базе данных
+            HomePage homePage1 = new HomePage();
+            homePage1.setBouquet(bouquet1);
+            homePage1.setDescription(description1);
+            homePageRepository.save(homePage1);
 
-        HomePage homePage2 = new HomePage();
-        homePage2.setBouquet(bouquet2);
-        homePage2.setDescription(description2);
-        homePageRepository.save(homePage2);
+            HomePage homePage2 = new HomePage();
+            homePage2.setBouquet(bouquet2);
+            homePage2.setDescription(description2);
+            homePageRepository.save(homePage2);
 
-        HomePage homePage3 = new HomePage();
-        homePage3.setBouquet(bouquet3);
-        homePage3.setDescription(description3);
-        homePageRepository.save(homePage3);
+            HomePage homePage3 = new HomePage();
+            homePage3.setBouquet(bouquet3);
+            homePage3.setDescription(description3);
+            homePageRepository.save(homePage3);
 
-        // Вернуть страницу главного экрана
-        return "redirect:/";
+            return "redirect:/";
+
+        } else {
+            redirectAttributes.addFlashAttribute("error", "Вы заполнили не все данные. Операция отклонена.");
+            return "redirect:/admin";
+        }
+
     }
 }
